requirements.txt
google-generativeai
python-dotenv
selenium
python-telegram-bot 